+++ argilla-frontend/v1/domain/entities/dataset/Metrics.ts
@@ -6,6 +6,9 @@ export class Metrics {
     discarded: number;
   };
 
+  public readonly hasMetrics: boolean;
+  public readonly isEmpty: boolean;
+
   constructor(
     public readonly total: number,
     public readonly submitted: number,
@@ -13,16 +16,15 @@ export class Metrics {
     public readonly draft: number,
     public readonly pending: number
   ) {
+    this.hasMetrics = total >= 0;
+    this.isEmpty = total === 0;
+
     this.percentage = {
       pending: (this.pending * 100) / this.total,
       draft: (this.draft * 100) / this.total,
       submitted: (this.submitted * 100) / this.total,
       discarded: (this.discarded * 100) / this.total,
     };
-  }
-
-  get hasMetrics() {
-    return this.total > 0;
   }
 
   get responded() {