+++ argilla-server/src/argilla_server/_app.py
@@ -19,6 +19,7 @@ import logging
 import os
 import shutil
 import tempfile
+import redis
 from datetime import datetime
 from pathlib import Path
 
@@ -40,6 +41,8 @@ from argilla_server.models import User, Workspace
 from argilla_server.search_engine import get_search_engine
 from argilla_server.settings import settings
 from argilla_server.static_rewrite import RewriteStaticFiles
+from argilla_server.jobs.queues import REDIS_CONNECTION
+from argilla_server.telemetry import get_telemetry_client
 
 _LOGGER = logging.getLogger("argilla")
 
@@ -47,9 +50,11 @@ _LOGGER = logging.getLogger("argilla")
 @contextlib.asynccontextmanager
 async def app_lifespan(app: FastAPI):
     # See https://fastapi.tiangolo.com/advanced/events/#lifespan
-    show_telemetry_warning()
     await configure_database()
     await configure_search_engine()
+    configure_redis()
+    track_server_startup()
+
     yield
 
 
@@ -67,8 +72,9 @@ def create_server_app() -> FastAPI:
     )
 
     configure_logging()
-    configure_middleware(app)
+    configure_common_middleware(app)
     configure_api_router(app)
+    configure_telemetry(app)
     configure_app_statics(app)
     configure_api_docs(app)
 
@@ -92,7 +98,7 @@ def configure_api_docs(app: FastAPI):
         return RedirectResponse(url=f"{settings.base_url}api/v1/docs")
 
 
-def configure_middleware(app: FastAPI):
+def configure_common_middleware(app: FastAPI):
     """Configures fastapi middleware"""
 
     @app.middleware("http")
@@ -119,6 +125,24 @@ def configure_middleware(app: FastAPI):
 def configure_api_router(app: FastAPI):
     """Configures and set the api router to app"""
     app.mount("/api/v1", api_v1)
+
+
+def configure_telemetry(app: FastAPI):
+    """
+    Configures telemetry middleware for the app if telemetry is enabled
+    """
+    if not settings.enable_telemetry:
+        return
+
+    @app.middleware("http")
+    async def track_api_requests(request: Request, call_next):
+        response = await call_next(request)
+        try:
+            await get_telemetry_client().track_api_request(request, response)
+        except Exception as e:
+            _LOGGER.warning(f"Error tracking request: {e}")
+        finally:
+            return response
 
 
 def configure_app_statics(app: FastAPI):
@@ -165,18 +189,28 @@ def configure_app_statics(app: FastAPI):
     )
 
 
-def show_telemetry_warning():
-    if settings.enable_telemetry:
-        message = "\n"
-        message += inspect.cleandoc(
-            "Argilla uses telemetry to report anonymous usage and error information. You\n"
-            "can know more about what information is reported at:\n\n"
-            "    https://docs.argilla.io/latest/reference/argilla-server/telemetry/\n\n"
-            "Telemetry is currently enabled. If you want to disable it, you can configure\n"
-            "the environment variable before relaunching the server:\n\n"
-            f'{"#set ARGILLA_ENABLE_TELEMETRY=0" if os.name == "nt" else "$>export ARGILLA_ENABLE_TELEMETRY=0"}'
-        )
-        _LOGGER.warning(message)
+def track_server_startup() -> None:
+    """
+    Track server startup telemetry event if telemetry is enabled
+    """
+    if not settings.enable_telemetry:
+        return
+
+    _show_telemetry_warning()
+    get_telemetry_client().track_server_startup()
+
+
+def _show_telemetry_warning():
+    message = "\n"
+    message += inspect.cleandoc(
+        "Argilla uses telemetry to report anonymous usage and error information. You\n"
+        "can know more about what information is reported at:\n\n"
+        "    https://docs.argilla.io/latest/reference/argilla-server/telemetry/\n\n"
+        "Telemetry is currently enabled. If you want to disable it, you can configure\n"
+        "the environment variable before relaunching the server:\n\n"
+        f'{"#set HF_HUB_DISABLE_TELEMETRY=1" if os.name == "nt" else "$>export HF_HUB_DISABLE_TELEMETRY=1"}'
+    )
+    _LOGGER.warning(message)
 
 
 async def _create_oauth_allowed_workspaces(db: AsyncSession):
@@ -233,6 +267,22 @@ async def configure_search_engine():
                 )
 
     await ping_search_engine()
+
+
+def configure_redis():
+    @backoff.on_exception(backoff.expo, ConnectionError, max_time=60)
+    def ping_redis():
+        try:
+            REDIS_CONNECTION.ping()
+        except redis.exceptions.ConnectionError:
+            raise ConnectionError(
+                f"Your redis instance at {settings.redis_url} is not available or not responding.\n"
+                "Please make sure your redis instance is launched and correctly running and\n"
+                "you have the necessary access permissions. Once you have verified this, restart "
+                "the argilla server.\n"
+            )
+
+    ping_redis()
 
 
 app = create_server_app()