+++ argilla-server/src/argilla_server/validators/records.py
@@ -91,20 +91,6 @@ class RecordValidatorBase(ABC):
                     "and extra metadata is not allowed for this dataset"
                 )
 
-    @staticmethod
-    def _validate_required_fields(dataset: Dataset, fields: Dict[str, str]) -> None:
-        for field in dataset.fields:
-            if field.required and not (field.name in fields and fields.get(field.name) is not None):
-                raise UnprocessableEntityError(f"missing required value for field: {field.name!r}")
-
-    @staticmethod
-    def _validate_extra_fields(dataset: Dataset, fields: Dict[str, str]) -> None:
-        fields_copy = copy.copy(fields)
-        for field in dataset.fields:
-            fields_copy.pop(field.name, None)
-        if fields_copy:
-            raise UnprocessableEntityError(f"found fields values for non configured fields: {list(fields_copy.keys())}")
-
     @classmethod
     def _validate_image_fields(cls, dataset: Dataset, fields: Dict[str, str]) -> None:
         for field in filter(lambda field: field.is_image, dataset.fields):