+++ argilla/src/argilla/_models/_settings/_fields.py
@@ -36,7 +36,11 @@ class ChatFieldSettings(BaseModel):
 
 
 FieldSettings = Annotated[
-    Union[TextFieldSettings, ImageFieldSettings, ChatFieldSettings],
+    Union[
+        TextFieldSettings,
+        ImageFieldSettings,
+        ChatFieldSettings,
+    ],
     Field(..., discriminator="type"),
 ]