+++ argilla/src/argilla/datasets/_export/_hub.py
@@ -16,11 +16,15 @@ import os
 import warnings
 from collections import defaultdict
 from tempfile import TemporaryDirectory
-from typing import TYPE_CHECKING, Any, Optional, Type, Union
+from typing import TYPE_CHECKING, Any, Optional, Type, Union, Dict
 from uuid import UUID
 
+from datasets import DatasetDict
 from datasets.data_files import EmptyDatasetError
 
+from argilla._exceptions._api import UnprocessableEntityError
+from argilla._exceptions._records import RecordsIngestionError
+from argilla._exceptions._settings import SettingsError
 from argilla.datasets._export._disk import DiskImportExportMixin
 from argilla.records._mapping import IngestedRecordMapper
 from argilla.responses import Response
@@ -28,7 +32,7 @@ from argilla.responses import Response
 if TYPE_CHECKING:
     from datasets import Dataset as HFDataset
 
-    from argilla import Argilla, Dataset, Workspace
+    from argilla import Argilla, Dataset, Workspace, Settings
 
 
 class HubImportExportMixin(DiskImportExportMixin):
@@ -69,6 +73,7 @@ class HubImportExportMixin(DiskImportExportMixin):
 
         with TemporaryDirectory() as tmpdirname:
             config_dir = os.path.join(tmpdirname)
+
             self.to_disk(path=config_dir, with_records=False)
 
             if generate_card:
@@ -110,6 +115,7 @@ class HubImportExportMixin(DiskImportExportMixin):
         workspace: Optional[Union["Workspace", str]] = None,
         client: Optional["Argilla"] = None,
         with_records: bool = True,
+        settings: Optional["Settings"] = None,
         **kwargs: Any,
     ):
         """Loads a `Dataset` from the Hugging Face Hub.
@@ -125,32 +131,32 @@ class HubImportExportMixin(DiskImportExportMixin):
         Returns:
             A `Dataset` loaded from the Hugging Face Hub.
         """
-        from datasets import Dataset, DatasetDict, load_dataset
+        from datasets import load_dataset
         from huggingface_hub import snapshot_download
 
-        # download both files in parallel
-        folder_path = snapshot_download(
-            repo_id=repo_id,
-            repo_type="dataset",
-            allow_patterns=cls._DEFAULT_CONFIGURATION_FILES,
-            token=kwargs.get("token"),
-        )
+        if name is None:
+            name = repo_id.replace("/", "_")
 
-        dataset = cls.from_disk(
-            path=folder_path, workspace=workspace, name=name, client=client, with_records=with_records
-        )
+        if settings is not None:
+            dataset = cls(name=name, settings=settings)
+            dataset.create()
+        else:
+            # download configuration files from the hub
+            folder_path = snapshot_download(
+                repo_id=repo_id,
+                repo_type="dataset",
+                allow_patterns=cls._DEFAULT_CONFIGURATION_FILES,
+                token=kwargs.get("token"),
+            )
+
+            dataset = cls.from_disk(
+                path=folder_path, workspace=workspace, name=name, client=client, with_records=with_records
+            )
 
         if with_records:
             try:
-                hf_dataset: Dataset = load_dataset(path=repo_id, **kwargs)  # type: ignore
-                if isinstance(hf_dataset, DatasetDict) and "split" not in kwargs:
-                    if len(hf_dataset.keys()) > 1:
-                        raise ValueError(
-                            "Only one dataset can be loaded at a time, use `split` to select a split, available splits"
-                            f" are: {', '.join(hf_dataset.keys())}."
-                        )
-                    hf_dataset: Dataset = hf_dataset[list(hf_dataset.keys())[0]]
-
+                hf_dataset = load_dataset(path=repo_id, **kwargs)  # type: ignore
+                hf_dataset = cls._get_dataset_split(hf_dataset=hf_dataset, **kwargs)
                 cls._log_dataset_records(hf_dataset=hf_dataset, dataset=dataset)
             except EmptyDatasetError:
                 warnings.warn(
@@ -198,9 +204,7 @@ class HubImportExportMixin(DiskImportExportMixin):
         records = []
         for idx, row in enumerate(hf_dataset):
             record = mapper(row)
-            record.id = row.pop("id")
             for question_name, values in response_questions.items():
-                response_users = {}
                 response_values = values["responses"][idx]
                 response_users = values["users"][idx]
                 response_status = values["status"][idx]
@@ -217,4 +221,31 @@ class HubImportExportMixin(DiskImportExportMixin):
                     )
                     record.responses.add(response)
             records.append(record)
-        dataset.records.log(records=records)
+
+        try:
+            dataset.records.log(records=records)
+        except (RecordsIngestionError, UnprocessableEntityError) as e:
+            raise SettingsError(
+                message=f"Failed to load records from Hugging Face dataset. Defined settings do not match dataset schema. Hugging face dataset features: {hf_dataset.features}. Argilla dataset settings : {dataset.settings}"
+            ) from e
+
+    @staticmethod
+    def _get_dataset_split(hf_dataset: "HFDataset", split: Optional[str] = None, **kwargs: Dict) -> "HFDataset":
+        """Get a single dataset from a Hugging Face dataset.
+
+        Parameters:
+            hf_dataset (HFDataset): The Hugging Face dataset to get a single dataset from.
+
+        Returns:
+            HFDataset: The single dataset.
+        """
+
+        if isinstance(hf_dataset, DatasetDict) and split is None:
+            split = next(iter(hf_dataset.keys()))
+            if len(hf_dataset.keys()) > 1:
+                warnings.warn(
+                    message=f"Multiple splits found in Hugging Face dataset. Using the first split: {split}. "
+                    f"Available splits are: {', '.join(hf_dataset.keys())}."
+                )
+            hf_dataset = hf_dataset[split]
+        return hf_dataset